flask-backend/requirements.txt

Flask==2.3.3
Flask-CORS==4.0.0
vaderSentiment==3.3.2
supabase==1.0.4
PyJWT==2.8.0
python-dotenv==1.0.0
gunicorn==21.2.0
